functions/fileops-FileConverter/requirements.txt
aws-psycopg2==1.3.8
chardet==5.2.0
install==1.3.5
psycopg2-binary==2.9.7
simplejson==3.19.2
s3fs==2024.2.0
pandas==2.2.1
boto3==1.34.51
botocore==1.34.51
pyarrow==15.0.0